Split out select visitors tests
